requirements.txt
django
djangorestframework
psycopg2
pillow
django-environ